Move wallets/metadata API under customers
