+++ app/controllers/api/v1/wallets/metadata_controller.rb
@@ -4,56 +4,22 @@ module Api
   module V1
     module Wallets
       class MetadataController < BaseController
-        def create
-          result = ::Wallets::UpdateService.call(wallet:, params: metadata_params)
+        include WalletMetadataActions
 
-          if result.success?
-            render_metadata
-          else
-            render_error_response(result)
-          end
+        def create
+          metadata_create(wallet)
         end
 
         def update
-          result = ::Wallets::UpdateService.call(wallet:, partial_metadata: true, params: metadata_params)
-
-          if result.success?
-            render_metadata
-          else
-            render_error_response(result)
-          end
+          metadata_update(wallet)
         end
 
         def destroy
-          result = ::Wallets::UpdateService.call(wallet:, params: {metadata: nil})
-
-          if result.success?
-            render_metadata
-          else
-            render_error_response(result)
-          end
+          metadata_destroy(wallet)
         end
 
         def destroy_key
-          return not_found_error(resource: "metadata") unless wallet.metadata
-
-          result = Metadata::DeleteItemKeyService.call(item: wallet.metadata, key: params[:key])
-
-          if result.success?
-            render_metadata
-          else
-            render_error_response(result)
-          end
-        end
-
-        private
-
-        def metadata_params
-          params.permit(metadata: {}).to_h
-        end
-
-        def render_metadata
-          render(json: {metadata: wallet.reload.metadata&.value})
+          metadata_destroy_key(wallet)
         end
       end
     end
